requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
cryptography==41.0.5
web3==6.11.0 